Add a working nix flake (#643)
This pull request adds a proper nix derivation for `kmir`. The
derivation is build in the CI for testing and pushed to cachix during
release.

The `uv` version used by CI and docker is now always in sync with the
version used by nix. Having them use a different version resulted in
incompatible lock files in the past. You can enter a nix develop shell
with the proper `uv` version provided by the shell by running `nix
develop`.

Also, you can build and use the derivation `kmir.rust` instead of
`kmir`, in order to provide `kmir` with a proper rust distribution at
runtime that matches the rust distribution used to build
`stable-mir-json`.