Compiling SMIR JSON to kore to avoid static data in config (#745)
Pervasive changes to the semantics and python client code to replace the
lookup tables in the config (`memory`, `functions`, `types`) by lookup
functions whose equations are generated from the `SMIRInfo` i.e., from
the SMIR JSON , for each program.

Previously we had `Map`-sorted config cells `functions`, `types` and
`memory`, which are now _functions_ instead. The function equations have
to be added and then compiled with the LLVM backend for it to work, but
the configuration is much smaller and does not have any static data any
more.