Add `pyproject-nix` overlay in nix derivation (#724)
Previously, when RV nix derivations were using `poetry2nix`, overriding
a nix input in nix was also overriding the python code and resources
used in `poetry2nix`. This allowed for easy overrides of python and K
semantics changes by utilizing nix overrides.

With the migration to `uv` and `uv2nix`, this was behavior was removed
due to the respective code not documenting it's purpose. In addition,
due to `uv2nix` relying on state-of-the-art python nix packaging tooling
`pyproject-nix` instead of `nixpkgs`, ensuring identical override
behaviour requires more nix logic.

This PR re-introduces this kind of behaviour with `uv2nix`, though it
depends on newly introduced overlays in the k nix flake.

### Example
You can override k by running: `kup install kmir --override k-framework
~/path/to/local-k-checkout`. This will now also build kmir with the
python and K code in `~/path/to/local-k-checkout`.

---------

Co-authored-by: Jost Berthold <jost.berthold@gmail.com>
Co-authored-by: Tamás Tóth <tothtamas28@users.noreply.github.com>