Specialize tests by multisig config flag (#74)
This PR removes the configuration flags `#[cfg(feature="multisig")]` and
`#[cfg(not(feature="multisig"))]` from
`entrypoint-runtime-verification.rs`. To this end, tests with these
configuration flags are split into two, one for each configuration case.
In particular,
- Specialize tests dealing with an authority account depending on
whether the authority is a multisig or a regular account
- Adapt the dispatcher functions `inner_process_instruction` and
`inner_process_remaining_instruction` to accommodate for test
specialization
- Update `proofs.md` with the test changes
  - Update test names in the existing table
- Create a new table with an extra column for the multisig versions of
the tests
- Add `-m` flag to the `run-proofs.sh` to run the proofs in the multisig
table

---------

Co-authored-by: Jost Berthold <jost.berthold@gmail.com>