Fix offset deref for single elements (#798)
Faulty behaviour when dereferencing pointers after offset was observed
in the P-token proofs, returning an array instead of a single element.
The quick-fix is to project to a single element during `#derefTruncate`
when metadata indicates that a single value is required.

Also adds offset tests to `exec_smir` suite. Closes #764