NO OP rules for compiler hint intrinsics (#828)
Some instrinsics do not affect the semantics of the compiled program but
are hints to the compiler. In our semantics these intrinsics are NO OPs
where the call is ignored and current continuation (`<k>` cell) moves
onto the next term in the sequence.

In particular there are rules implemented for these intrinsics:
- `cold_path`
- `prefetch_read_data`
- `prefetch_write_data`
- `prefetch_read_instruction`
- `prefetch_write_instruction`

By handling `cold_path` we also support `likely` and `unlikely` which
are `MonoItemFn` that call the `cold_path` instrinic.

This PR also refactors the intrinsics into their only file
`intrinsics.md`.

closes #802

---------

Co-authored-by: Jost Berthold <jost.berthold@gmail.com>