`--terminate-on-thunk` option to stop execution on `thunk` (#834)
This PR adds `--terminate-on-thunk` option to the `kmir` cli. Enabling
this option will mean that execution will terminate if a `thunk` is
applied.

The technical details of how this applied is a cut point rule for
`thunk` is applied which means `thunk` will be at the top of the `<k>`
cell. Then the terminal rule is able to match that `thunk`.

---------

Co-authored-by: Tamás Tóth <tothtamas28@users.noreply.github.com>