Union Value Sort (#861)
This PR separates a `Value` sort for `Union` out of `Aggregate` as there
are different semantics for the projections (and it also makes the code
clearer). A `Union` has one argument which is data, but it can be
interpreted as different types via field projections - this is
significantly different from other field projections that access
different data via different arguments.

Currently the implementation only supports accessing fields of the same
type as the `Union` construction (no switching between types right now).