Tweak proofs to avoid vacuous and overflow branches (#122)
Overflow checks in the property proofs are converted to using
`x.checked_add(y)` instead of a custom predicate `u64::MAX - y < x` to
ensure the SMT solver can handle them without having to add
simplification lemmas.

Some of the proofs rely on the assumption that the total supply of
tokens is bound by `u64` and cannot be exceeded by any transfer or other
operation moving tokens from one account to another. Cases where this
would lead to overflow are skipped. This is protected by the
`assumptions` feature flag in the code.

This makes a number of proofs pass that previously exhibited
`assert_failed` and overflow errors.