Special fix-up rule to apply field projection on array head elements (#876)
When using pointer offsets followed by dereferencing, the returned
result is an array (subslice of the original).
In some cases, the value is used as a single element, for instance by
projecting out a field. This PR adds a special rule for the field
projection case to proceed using the head element in those cases.

Related: #771